back/requirements.txt
fastapi
numpy
keras
tensorflow
uvicorn
python-multipart
gunicorn
tensorflow_hub
Pillow